Move setup methods to func test class (#196)
