+++ pyEpiabm/pyEpiabm/tests/test_func/test_closure_functional.py
@@ -1,6 +1,6 @@
 import pandas as pd
 import unittest
-from unittest.mock import patch, mock_open, Mock
+from unittest.mock import patch, Mock
 
 import pyEpiabm as pe
 from pyEpiabm.tests import TestFunctional
@@ -13,19 +13,7 @@ class TestClosureFunctional(TestFunctional):
     results/properties to ensure code functions as desired.
     """
     def setUp(self) -> None:
-        self.pop_params = {'cell': [1.0, 2.0], 'microcell': [1.0, 1.0],
-                           'location_x': [0.0, 1.0], 'location_y': [0.0, 1.0],
-                           'household_number': [1, 1],
-                           'Susceptible': [80, 90], 'InfectMild': [10, 0],
-                           'place_number': 6}
-        self.sim_params = {"simulation_start_time": 0,
-                           "simulation_end_time": 15,
-                           "initial_infected_number": 0}
-
-        self.file_params = {"output_file": "output.csv",
-                            "output_dir": "test_folder/integration_tests",
-                            "spatial_output": False,
-                            "age_stratified": True}
+        TestFunctional.setUpPopulation()
 
         self.intervention = {"place_closure": {
             "start_time": 0,
@@ -39,34 +27,6 @@ class TestClosureFunctional(TestFunctional):
             "case_microcell_threshold": 0
         }
         }
-
-
-    @classmethod
-    def file_simulation(self, pop_file, sim_params, file_params, sweep_list):
-        # Create a population based on the parameters given.
-        population = pe.routine.FilePopulationFactory.make_pop(
-            pop_file, random_seed=42)
-        pe.routine.FilePopulationFactory.print_population(population,
-                                                          "test.csv")
-
-        mo = mock_open()
-        with patch('pyEpiabm.output._csv_dict_writer.open', mo):
-            sim = pe.routine.Simulation()
-            sim.configure(
-                population,
-                [pe.sweep.InitialInfectedSweep(),
-                 pe.sweep.InitialisePlaceSweep(),
-                 pe.sweep.InitialHouseholdSweep()],
-                sweep_list,
-                sim_params,
-                file_params)
-
-            sim.run_sweeps()
-
-        # Need to close the writer object at the end of each simulation.
-        del sim.writer
-        del sim
-        return population
 
     @patch('pyEpiabm.routine.simulation.tqdm', TestFunctional.notqdm)
     @patch('pyEpiabm.output._CsvDictWriter.write', Mock())
@@ -84,13 +44,13 @@ class TestClosureFunctional(TestFunctional):
         pe.Parameters.instance().infection_radius = 1.6
 
         # Without intervention
-        pop = TestClosureFunctional.file_simulation(
+        pop = TestFunctional.file_simulation(
             "test_input.csv", self.sim_params, self.file_params,
             HelperFunc.sweep_list_initialise()[1:])
 
         # Enable place closure
         pe.Parameters.instance().intervention_params = self.intervention
-        pop_closure = TestClosureFunctional.file_simulation(
+        pop_closure = TestFunctional.file_simulation(
             "test_input.csv", self.sim_params, self.file_params,
             HelperFunc.sweep_list_initialise())
 
@@ -118,13 +78,13 @@ class TestClosureFunctional(TestFunctional):
         pe.Parameters.instance().infection_radius = 1.6
 
         # Without intervention
-        pop = TestClosureFunctional.file_simulation(
+        pop = TestFunctional.file_simulation(
             "test_input.csv", self.sim_params, self.file_params,
             HelperFunc.sweep_list_initialise()[1:])
 
         pe.Parameters.instance().intervention_params = self.intervention
         self.intervention['place_closure']['closure_place_type'] = []
-        pop_closure = TestClosureFunctional.file_simulation(
+        pop_closure = TestFunctional.file_simulation(
             "test_input.csv", self.sim_params, self.file_params,
             HelperFunc.sweep_list_initialise())
 
@@ -151,13 +111,13 @@ class TestClosureFunctional(TestFunctional):
         pe.Parameters.instance().infection_radius = 1.6
 
         pe.Parameters.instance().intervention_params = self.intervention
-        pop_standard = TestClosureFunctional.file_simulation(
+        pop_standard = TestFunctional.file_simulation(
             "test_input.csv", self.sim_params, self.file_params,
             HelperFunc.sweep_list_initialise())
 
         self.intervention['place_closure']['closure_place_type'] = [
             1, 2, 3, 4, 5, 6]
-        pop = TestClosureFunctional.file_simulation(
+        pop = TestFunctional.file_simulation(
             "test_input.csv", self.sim_params, self.file_params,
             HelperFunc.sweep_list_initialise())
 
@@ -185,13 +145,13 @@ class TestClosureFunctional(TestFunctional):
         pe.Parameters.instance().infection_radius = 1.6
 
         pe.Parameters.instance().intervention_params = self.intervention
-        pop_standard = TestClosureFunctional.file_simulation(
+        pop_standard = TestFunctional.file_simulation(
             "test_input.csv", self.sim_params, self.file_params,
             HelperFunc.sweep_list_initialise())
 
         self.intervention['place_closure'][
             'closure_spatial_params'] = 1
-        pop = TestClosureFunctional.file_simulation(
+        pop = TestFunctional.file_simulation(
             "test_input.csv", self.sim_params, self.file_params,
             HelperFunc.sweep_list_initialise())
 
@@ -220,14 +180,14 @@ class TestClosureFunctional(TestFunctional):
         pe.Parameters.instance().infection_radius = 1.6
 
         # Without intervention
-        pop = TestClosureFunctional.file_simulation(
+        pop = TestFunctional.file_simulation(
             "test_input.csv", self.sim_params, self.file_params,
             HelperFunc.sweep_list_initialise()[1:])
 
         pe.Parameters.instance().intervention_params = self.intervention
         self.intervention['place_closure'][
             'case_microcell_threshold'] = 1000
-        pop_closure = TestClosureFunctional.file_simulation(
+        pop_closure = TestFunctional.file_simulation(
             "test_input.csv", self.sim_params, self.file_params,
             HelperFunc.sweep_list_initialise())
 
@@ -254,13 +214,13 @@ class TestClosureFunctional(TestFunctional):
         pe.Parameters.instance().infection_radius = 1.6
 
         pe.Parameters.instance().intervention_params = self.intervention
-        pop_standard = TestClosureFunctional.file_simulation(
+        pop_standard = TestFunctional.file_simulation(
             "test_input.csv", self.sim_params, self.file_params,
             HelperFunc.sweep_list_initialise())
 
         self.intervention['place_closure'][
             'case_microcell_threshold'] = 15
-        pop = TestClosureFunctional.file_simulation(
+        pop = TestFunctional.file_simulation(
             "test_input.csv", self.sim_params, self.file_params,
             HelperFunc.sweep_list_initialise())
