+++ pyEpiabm/pyEpiabm/tests/test_func/test_distancing_functional.py
@@ -1,7 +1,6 @@
-import os
 import pandas as pd
 import unittest
-from unittest.mock import patch, mock_open, Mock
+from unittest.mock import patch, Mock
 
 import pyEpiabm as pe
 from pyEpiabm.tests import TestFunctional
@@ -15,19 +14,7 @@ class TestDistancingFunctional(TestFunctional):
     """
 
     def setUp(self) -> None:
-        self.pop_params = {'cell': [1.0, 2.0], 'microcell': [1.0, 1.0],
-                           'location_x': [0.0, 1.0], 'location_y': [0.0, 1.0],
-                           'household_number': [1, 1],
-                           'Susceptible': [800, 900], 'InfectMild': [10, 0],
-                           'place_number': 6}
-        self.sim_params = {"simulation_start_time": 0,
-                           "simulation_end_time": 15,
-                           "initial_infected_number": 0}
-
-        self.file_params = {"output_file": "output.csv",
-                            "output_dir": "test_folder/integration_tests",
-                            "spatial_output": False,
-                            "age_stratified": True}
+        TestFunctional.setUpPopulation()
 
         self.intervention = {"social_distancing": {
             "start_time": 0,
@@ -47,32 +34,6 @@ class TestDistancingFunctional(TestFunctional):
         }
         }
 
-    def file_simulation(pop_file, sim_params, file_params, sweep_list):
-        # Create a population based on the parameters given.
-        population = pe.routine.FilePopulationFactory.make_pop(
-            pop_file, random_seed=42)
-        pe.routine.FilePopulationFactory.print_population(population,
-                                                          "test.csv")
-
-        mo = mock_open()
-        with patch('pyEpiabm.output._csv_dict_writer.open', mo):
-            sim = pe.routine.Simulation()
-            sim.configure(
-                population,
-                [pe.sweep.InitialInfectedSweep(),
-                 pe.sweep.InitialisePlaceSweep(),
-                 pe.sweep.InitialHouseholdSweep()],
-                sweep_list,
-                sim_params,
-                file_params)
-
-            sim.run_sweeps()
-
-        # Need to close the writer object at the end of each simulation.
-        del sim.writer
-        del sim
-        return population
-
     @patch('pyEpiabm.routine.simulation.tqdm', TestFunctional.notqdm)
     @patch('pyEpiabm.output._CsvDictWriter.write', Mock())
     @patch('os.makedirs', Mock())
@@ -89,13 +50,13 @@ class TestDistancingFunctional(TestFunctional):
         pe.Parameters.instance().infection_radius = 1.6
 
         # Without intervention
-        pop = TestDistancingFunctional.file_simulation(
+        pop = TestFunctional.file_simulation(
             "test_input.csv", self.sim_params, self.file_params,
             HelperFunc.sweep_list_initialise()[1:])
 
         # Enable social distancing
         pe.Parameters.instance().intervention_params = self.intervention
-        pop_distancing = TestDistancingFunctional.file_simulation(
+        pop_distancing = TestFunctional.file_simulation(
             "test_input.csv", self.sim_params, self.file_params,
             HelperFunc.sweep_list_initialise())
 
@@ -122,13 +83,13 @@ class TestDistancingFunctional(TestFunctional):
         pe.Parameters.instance().infection_radius = 1.6
 
         pe.Parameters.instance().intervention_params = self.intervention
-        pop_standard = TestDistancingFunctional.file_simulation(
+        pop_standard = TestFunctional.file_simulation(
             "test_input.csv", self.sim_params, self.file_params,
             HelperFunc.sweep_list_initialise())
 
         self.intervention['social_distancing'][
             'distancing_spatial_enhanced_susc'] = 0.8
-        pop = TestDistancingFunctional.file_simulation(
+        pop = TestFunctional.file_simulation(
             "test_input.csv", self.sim_params, self.file_params,
             HelperFunc.sweep_list_initialise())
 
@@ -156,13 +117,13 @@ class TestDistancingFunctional(TestFunctional):
         pe.Parameters.instance().infection_radius = 1.6
 
         pe.Parameters.instance().intervention_params = self.intervention
-        pop_standard = TestDistancingFunctional.file_simulation(
+        pop_standard = TestFunctional.file_simulation(
             "test_input.csv", self.sim_params, self.file_params,
             HelperFunc.sweep_list_initialise())
 
         self.intervention['social_distancing'][
             'distancing_enhanced_prob'] = [0.1]*17
-        pop = TestDistancingFunctional.file_simulation(
+        pop = TestFunctional.file_simulation(
             "test_input.csv", self.sim_params, self.file_params,
             HelperFunc.sweep_list_initialise())
